Harden script running
When running a script with a lot of data coming back, the script will
write multiple times. We handled only one response before but
needed to loop to handle more.

We are also now handling bad exit codes instead of crashing on a bad
return.

I was looking into using a two-arity function as an Enumerable for the
processing, but it is poorly documented, and I thought we could deal
with it at another junction.

@akoutmos pointed out a beautiful way to handle looping with
`String.cycle([nil])`. I decided to mimic that code and took it almost
entirely from https://github.com/akoutmos/mjml_eex/blob/master/lib/compilers/node.ex#L72-L98

Thank you, @akoutmos, for that pointer.