Installs Deno as part of the build process
To ensure that deno is installed and reduce the users' overhead, we are
installing deno as part of the build process. Deno will only be
installed if it doesn't already exist in the configured location.

I increased the default timeout because running `mix test` would timeout
due to the upfront cost of Deno setting up its environment. I decided on
5000 milliseconds as the new timeout because it is long enough for any
system and seems a standard default for timeouts in Elixir.