Makes error reporting more reliable
We reported anything with `STDERR` output as an error status. This isn't
true, and some scripts may write to `STDERR` even if the status is still
good. Now we base the status on the way the process exists. We have
three types of return statuses from `DenoEx.run/4`. It may return
`{:ok, stdout_contnet}`, `{:error, stderr_contnet}`, or
`{:timeout, #DenoEx.Pipe{}}`.

I broke out how a Pipe is handled into `DenoEx.Pipe`. This allowed
better handling of the exit statuses, and allows more flexibility for
the end user. The user now has two options. One is to use the existing
`DenoEx.run/4` will handle the pipe for you and return the results Use
`DenoEx.Pipe` of the script and handle everything yourself.

`DenoEx.Pipe` is better suited to deal with long-running processes and
asynchronous processing of the responses.