Adds Styler to the project
Styler fixes some of the credo checks that can be automated, so we don't
have to think about it. It also rearranges the file into a standard
order. This helps because while we are working, we can stay at the
same level as the code we are writing and add our imports and everything
there. Then when we format, it will move those to the correct locations
so we never have to think about it again.

There is one side effect. Since some of our module-attributes depended
on other module-attributes, i.e., documentation, we need to move those to
bare identifiers outside the module. These identifiers only exist
during the compilation process.

I believe this may help keep things consistent even with outside
contributors and without putting as much onus on them.