Fixes scripts exiting with bad codes
About 20% of the time when running deno scripts we would receive an
exit status of 1 which is `sighup`. It would happen when running the
same script over and over and seemed very random.

Attempting to use `Port` instead of `:exec` and we no longer have those
bad exits.

In order to get this to work we have to drop support for sending an eof
to the deno process. The only way to send an `eof` to a `Port` is to
close the `Port`, but that closes STDOUT as well. Since STDOUT is closed
we never get the response from the scripts.

This changed the way we had to handle scripts on STDIN. Now we utilize
`echo` to pass the file contents through a pipe to `deno run`.